Added a getValue function to the ClassField class which will return the value of the Field if initialized, else null.
Added a secondary constructor to ObjectHandler which takes in an Object
instead of a class name. The object is assumed to be initialized.

Added functionality to ObjectSerializerClient to first check if an
Object field has already been initialized. If not, it creates a new
instance of the object by name. Otherwise, it initializes ObjectHandler
with the already initialized object.